refactor(menubar): split layout planners into LayoutSolver and PendingLedger
The previous commit moved the four icon-restore orchestrators behind a set of pure planner functions but kept them grouped as an extension on `MenuBarItemManager`. The structure mixed snapshot-pure decision logic (cross-section rebalance, within-section reorder, notch overflow, LCS sequencing, full-sort sequencing, unmanaged placement, saved-section rebuild, anchor resolution, saved-position lookup) with one genuinely history-dependent planner (per-entry pending-relocation decision that consumes the waitForRelaunch sentinel and `pendingReturnDestinations` ledger state). The LLM council's "Path C" recommendation called for separating these two by temporality so future contributors do not need to re-discover which planners depend on retry state. This change physically realizes that split.

Adds `Thaw/MenuBar/MenuBarItems/LayoutSolver.swift` as a static-only `enum` containing the snapshot-pure planners and their result types: `planRebalanceMove`, `planLeftmostMove`, `planNotchOverflow`, `planLCSMoveSequence`, `planFullSortSequence`, `planWithinSectionReorder`, `planUnmanagedPlacement`, `planSectionOrder`, `anchorDestination`, `savedPosition`, `savedPositionByBaseID`, plus the `longestCommonSubsequence` helper. Result types `RebalanceMove`, `LeftmostMove`, `NotchOverflowResult`, `LCSPlannedMove`, `LCSPlannedDestination`, `WithinSectionMove`, `UnmanagedPlacement`, and `SavedPosition` move with them. Every function is `nonisolated static` and pure over its inputs; nothing inside the file calls `Bridging` or `NSScreen`.

Adds `Thaw/MenuBar/MenuBarItems/PendingLedger.swift` as the matching namespace for the history-dependent decision logic: `planPendingMove` along with `PendingEntry` and `PendingMove`. The orchestrator (`relocatePendingItems`) still owns the `pendingRelocations` and `pendingReturnDestinations` mutable state and the raw-string sentinel persistence; the ledger file owns the typed shapes and the per-entry decision algorithm.

Updates the four orchestrator call sites in `MenuBarItemManager.swift` to invoke `LayoutSolver.planXxx(...)` / `PendingLedger.planPendingMove(...)` instead of `Self.planXxx(...)`. Updates the 12 planner test files (`PlanRebalanceMoveTests`, `PlanLeftmostMoveTests`, `PlanPendingMoveTests`, `PlanNotchOverflowTests`, `PlanLCSMoveSequenceTests`, `PlanFullSortSequenceTests`, `PlanWithinSectionReorderTests`, `PlanUnmanagedPlacementTests`, `PlanSectionOrderTests`, `AnchorDestinationTests`, `SavedPositionLookupTests`, plus the existing `MenuBarTestFixtures*Tests`) to reference the new fully-qualified names. The four orchestrator functions remain in `MenuBarItemManager.swift` per the council's warning against collapsing them into one method.

`MenuBarItemManager.swift` shrinks from 7,770 lines to 6,567 lines (net `-1,203`). `LayoutSolver.swift` is 1,099 lines; `PendingLedger.swift` is 178 lines. Doc comments at the top of each new file explain the temporality boundary so future contributors understand which planner belongs where.

Pure refactor: no behavior change, no new tests, no removed tests. Full `ThawTests` suite remains at 626 pass, 0 fail. Public API of `MenuBarItemManager` is unchanged outside the planner namespace migration; existing callers of the four orchestrator methods continue to work without modification.
