+++ ThawTests/SavedPositionLookupTests.swift
@@ -20,11 +20,11 @@ final class SavedPositionLookupTests: XCTestCase {
             "visible": ["com.example.app:Status", "com.other.app:Item"],
             "hidden": ["com.example.app:Helper"],
         ]
-        let result = MenuBarItemManager.savedPosition(
+        let result = LayoutSolver.savedPosition(
             for: "com.other.app:Item",
             in: saved
         )
-        XCTAssertEqual(result, MenuBarItemManager.SavedPosition(section: .visible, index: 1))
+        XCTAssertEqual(result, LayoutSolver.SavedPosition(section: .visible, index: 1))
     }
 
     /// An identifier present in the hidden section returns .hidden.
@@ -33,11 +33,11 @@ final class SavedPositionLookupTests: XCTestCase {
             "visible": ["com.example.app:Status"],
             "hidden": ["com.example.app:Helper"],
         ]
-        let result = MenuBarItemManager.savedPosition(
+        let result = LayoutSolver.savedPosition(
             for: "com.example.app:Helper",
             in: saved
         )
-        XCTAssertEqual(result, MenuBarItemManager.SavedPosition(section: .hidden, index: 0))
+        XCTAssertEqual(result, LayoutSolver.SavedPosition(section: .hidden, index: 0))
     }
 
     /// An identifier not in any saved section returns nil.
@@ -45,7 +45,7 @@ final class SavedPositionLookupTests: XCTestCase {
         let saved: [String: [String]] = [
             "visible": ["com.example.app:Status"],
         ]
-        let result = MenuBarItemManager.savedPosition(
+        let result = LayoutSolver.savedPosition(
             for: "com.absent.app:Missing",
             in: saved
         )
@@ -54,7 +54,7 @@ final class SavedPositionLookupTests: XCTestCase {
 
     /// Empty savedSectionOrder returns nil.
     func testEmptySavedSectionOrder() {
-        let result = MenuBarItemManager.savedPosition(for: "anything", in: [:])
+        let result = LayoutSolver.savedPosition(for: "anything", in: [:])
         XCTAssertNil(result)
     }
 
@@ -64,11 +64,11 @@ final class SavedPositionLookupTests: XCTestCase {
         let saved: [String: [String]] = [
             "visible": ["com.example.app:Status", "com.example.app:Status:1", "com.example.app:Status:2"],
         ]
-        let result = MenuBarItemManager.savedPosition(
+        let result = LayoutSolver.savedPosition(
             for: "com.example.app:Status:1",
             in: saved
         )
-        XCTAssertEqual(result, MenuBarItemManager.SavedPosition(section: .visible, index: 1))
+        XCTAssertEqual(result, LayoutSolver.SavedPosition(section: .visible, index: 1))
     }
 
     // MARK: - savedPositionByBaseID (baseID fallback)
@@ -78,11 +78,11 @@ final class SavedPositionLookupTests: XCTestCase {
         let saved: [String: [String]] = [
             "visible": ["com.example.app:Status", "com.example.app:Status:1"],
         ]
-        let result = MenuBarItemManager.savedPositionByBaseID(
+        let result = LayoutSolver.savedPositionByBaseID(
             for: "com.example.app:Status:1",
             in: saved
         )
-        XCTAssertEqual(result, MenuBarItemManager.SavedPosition(section: .visible, index: 1),
+        XCTAssertEqual(result, LayoutSolver.SavedPosition(section: .visible, index: 1),
                        "exact :1 match should win even though :0 (no suffix) shares the baseID")
     }
 
@@ -95,11 +95,11 @@ final class SavedPositionLookupTests: XCTestCase {
         // A new instance shows up as :5 (e.g. spurious instanceIndex from
         // ordering churn). The exact match fails; baseID fallback returns
         // the first saved instance.
-        let result = MenuBarItemManager.savedPositionByBaseID(
+        let result = LayoutSolver.savedPositionByBaseID(
             for: "com.example.app:Status:5",
             in: saved
         )
-        XCTAssertEqual(result, MenuBarItemManager.SavedPosition(section: .hidden, index: 0),
+        XCTAssertEqual(result, LayoutSolver.SavedPosition(section: .hidden, index: 0),
                        "baseID fallback should return the first matching saved instance")
     }
 
@@ -108,7 +108,7 @@ final class SavedPositionLookupTests: XCTestCase {
         let saved: [String: [String]] = [
             "visible": ["com.example.app:Status"],
         ]
-        let result = MenuBarItemManager.savedPositionByBaseID(
+        let result = LayoutSolver.savedPositionByBaseID(
             for: "no-colon-here",
             in: saved
         )
